app/requirements.txt
sentence-transformers
fastapi
uvicorn[standard]
qdrant-client
langroid
fastembed
pymysql
python-dotenv
python-multipart
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
